remodel `Header` division
