+++ src/components/Global/Header/HeaderModule.tsx
@@ -1,6 +1,5 @@
 'use client'
 
-import LogoImage from '$/logo.svg'
 import {X} from 'lucide-react'
 
 import {contactsData} from '@/utils/contactsData'
@@ -9,21 +8,12 @@ import {useState, useEffect, useRef} from 'react'
 import {gsap} from 'gsap'
 import {useGSAP} from '@gsap/react'
 
-import Image from 'next/image'
 import {Link, Locale} from '@/i18n/routing'
 import Button from '~/UI/Button'
 import Text from '~/UI/Text'
-import HeaderLocale from '~/Global/HeaderLocale/HeaderLocale'
+import HeaderLocale from '~/Global/Header/HeaderLocale'
 
-const headerData = {
-  services: ['Услуги', 'Services'],
-  projects: ['Проекты', 'Projects'],
-  team: ['Команда', 'Team'],
-  news: ['Новости', 'News'],
-  contacts: ['Контакты', 'Contacts'],
-}
-
-export default function Header({locale}: {locale: Locale}) {
+export default function HeaderModule({locale, links, address}: {locale: Locale; links: {[key: string]: string[]}; address: string}) {
   const container = useRef<HTMLElement | null>(null)
   const [isMenuOpen, setIsMenuOpen] = useState(false)
 
@@ -75,13 +65,9 @@ export default function Header({locale}: {locale: Locale}) {
   }, [isMenuOpen])
 
   return (
-    <header className="fixed z-50 flex items-center justify-between w-screen sm:p-3 sm:bg-foreground bg-red">
-      <Link href="/" className="z-[99] px-4 py-3 sm:p-1.5 bg-red w-fit">
-        <Image quality={100} src={LogoImage} className="xl:w-[170px] sm:w-[110px] object-contain" alt="" />
-      </Link>
-
+    <>
       <nav className="flex items-center justify-around h-full gap-5 xl:gap-3.5 pr-8 sm:hidden bg-red">
-        {Object.entries(headerData).map(([key, value]) => (
+        {Object.entries(links).map(([key, value]) => (
           <Button to={`/#${key}`} className="flex-row-reverse py-2 xl:py-1.5 pl-11 pr-5 xl:pl-8 xl:pr-5 xl:w-full xl:gap-2 text-background xl:text-base" text={value[locale === 'ru' ? 0 : 1]} key={key} />
         ))}
 
@@ -101,7 +87,7 @@ export default function Header({locale}: {locale: Locale}) {
       <section ref={container} className={`absolute inset-0 z-20 w-screen h-screen ${isMenuOpen ? 'pointer-events-auto' : 'pointer-events-none'}`}>
         <div className="flex flex-col justify-between h-full px-4 pb-20 menu-overlay pt-28 bg-foreground text-background">
           <div className="flex flex-col gap-3.5">
-            {Object.entries(headerData).map(([key, value]) => (
+            {Object.entries(links).map(([key, value]) => (
               <Link href={`/#${key}`} className="text-3xl font-light uppercase menu-item w-fit" onClick={toggleMenu} key={key}>
                 {value[locale === 'ru' ? 0 : 1]}
               </Link>
@@ -112,7 +98,7 @@ export default function Header({locale}: {locale: Locale}) {
             <span className="font-extralight">Адрес</span>
 
             <Link href={contactsData.address.link}>
-              <Text type="h5" className="text-base font-medium !leading-[1.4]" text={contactsData.address.label} />
+              <Text type="h5" className="text-base font-medium !leading-[1.4]" text={address} />
             </Link>
           </div>
 
@@ -129,6 +115,6 @@ export default function Header({locale}: {locale: Locale}) {
           </div>
         </div>
       </section>
-    </header>
+    </>
   )
 }
